feat: improve home page design
Add the logo of the website and also use a different font for the welcome message.
Introduce a new picture with a specific style for showing a profile picture.
Also add a background image clipped.
Modify base layout to add generic footer and header.
Add header component with navigation.
Add footer component